Quix Game Commit
This a setbaclk edition of the quiz game for Assignemnt 2. I was working for hours trying to get it to be able to update the answers textboxes, and when I finally was able to assign them and run a test, Unity stalled and crashed. Upon reopening it, it had undid all the work on the UI I had preformed leaving me unable to be able to send an even patially working prototype video to you.